pyproject.toml
[project]
name = "sexkin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py-self",
    "openai>=2.32.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
]

[tool.uv.sources]
discord-py-self = { git = "https://github.com/dolfies/discord.py-self" }
